Add interaction.js for checkbox onchange; Ensure clear completed works
